host/fixer: Improve sirenia recovery
