+++ host/fixer/fixer.go
@@ -1,21 +1,15 @@
 package fixer
 
 import (
-	"encoding/json"
 	"fmt"
 	"sort"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/flynn/flynn/controller/client"
-	ct "github.com/flynn/flynn/controller/types"
-	"github.com/flynn/flynn/controller/utils"
 	"github.com/flynn/flynn/discoverd/client"
 	"github.com/flynn/flynn/host/types"
 	"github.com/flynn/flynn/pkg/cluster"
-	sirenia "github.com/flynn/flynn/pkg/sirenia/client"
-	state "github.com/flynn/flynn/pkg/sirenia/state"
 	"github.com/flynn/go-docopt"
 	"gopkg.in/inconshreveable/log15.v2"
 )
@@ -155,6 +149,15 @@ func (f *ClusterFixer) Run(args *docopt.Args, c *cluster.Client) error {
 	return nil
 }
 
+func (f *ClusterFixer) Host(id string) *cluster.Host {
+	for _, h := range f.hosts {
+		if h.ID() == id {
+			return h
+		}
+	}
+	return nil
+}
+
 func (f *ClusterFixer) StartAppJob(app, typ, service string) ([]*discoverd.Instance, error) {
 	f.l.Info(fmt.Sprintf("no %s %s process running, getting release details from hosts", app, typ))
 	releases := f.FindAppReleaseJobs(app, typ)
@@ -176,371 +179,6 @@ func (f *ClusterFixer) StartAppJob(app, typ, service string) ([]*discoverd.Insta
 	}
 	f.l.Info("waiting for job to start")
 	return discoverd.GetInstances(service, time.Minute)
-}
-
-func (f *ClusterFixer) FixDiscoverd() error {
-	f.l.Info("ensuring discoverd is running on all hosts")
-	releases := f.FindAppReleaseJobs("discoverd", "app")
-	if len(releases) == 0 {
-		return fmt.Errorf("didn't find any discoverd release jobs")
-	}
-outer:
-	for hostID, job := range releases[0] {
-		for _, h := range f.hosts {
-			if h.ID() != hostID {
-				continue
-			}
-
-			// check if discoverd is already running on this host
-			jobs, err := h.ListJobs()
-			if err != nil {
-				return fmt.Errorf("error listing jobs on %s: %s", h.ID(), err)
-			}
-			for _, j := range jobs {
-				if j.Status == host.StatusRunning &&
-					j.Job.Metadata["flynn-controller.app_name"] == "discoverd" &&
-					j.Job.Metadata["flynn-controller.type"] == "app" {
-					continue outer
-				}
-			}
-
-			job.ID = cluster.GenerateJobID(h.ID(), "")
-			f.FixJobEnv(job)
-			if err := h.AddJob(job); err != nil {
-				return fmt.Errorf("error starting discoverd on %s: %s", h.ID(), err)
-			}
-			f.l.Info("started discoverd instance", "job.id", job.ID)
-			break
-		}
-	}
-	return nil
-}
-
-func (f *ClusterFixer) FixFlannel() error {
-	f.l.Info("checking flannel")
-
-	flannelJobs := make(map[string]*host.Job, len(f.hosts))
-	for _, h := range f.hosts {
-		jobs, err := h.ListJobs()
-		if err != nil {
-			return fmt.Errorf("error getting jobs list from %s: %s", h.ID(), err)
-		}
-		for _, j := range jobs {
-			if j.Status != host.StatusRunning ||
-				j.Job.Metadata["flynn-controller.app_name"] != "flannel" ||
-				j.Job.Metadata["flynn-controller.type"] != "app" {
-				continue
-			}
-			flannelJobs[h.ID()] = j.Job
-			break
-		}
-	}
-	if len(flannelJobs) == len(f.hosts) {
-		f.l.Info("flannel looks good")
-		return nil
-	}
-
-	var job *host.Job
-	if len(flannelJobs) == 0 {
-		f.l.Info("flannel not running, starting it on each host")
-		releases := f.FindAppReleaseJobs("flannel", "app")
-		if len(releases) == 0 {
-			return fmt.Errorf("didn't find flannel release jobs")
-		}
-		for _, j := range releases[0] {
-			job = j
-			break
-		}
-	} else {
-		f.l.Info("flannel is not running on each host, starting missing jobs")
-		for _, job = range flannelJobs {
-			break
-		}
-	}
-
-	for _, h := range f.hosts {
-		if _, ok := flannelJobs[h.ID()]; ok {
-			continue
-		}
-		job.ID = cluster.GenerateJobID(h.ID(), "")
-		f.FixJobEnv(job)
-		if err := h.AddJob(job); err != nil {
-			return fmt.Errorf("error starting flannel job: %s", err)
-		}
-		f.l.Info("started flannel job", "job.id", job.ID)
-	}
-
-	f.l.Info("flannel fix complete")
-
-	return nil
-}
-
-func (f *ClusterFixer) FixController(instances []*discoverd.Instance, startScheduler bool) error {
-	f.l.Info("found controller instance, checking critical formations")
-	inst := instances[0]
-	client, err := controller.NewClient("http://"+inst.Addr, inst.Meta["AUTH_KEY"])
-	if err != nil {
-		return fmt.Errorf("unexpected error creating controller client: %s", err)
-	}
-
-	// check that formations for critical components are expected
-	apps := []string{"controller", "router", "discoverd", "flannel", "postgres"}
-	changes := make(map[string]*ct.Formation, len(apps))
-	var controllerFormation *ct.Formation
-	for _, app := range apps {
-		release, err := client.GetAppRelease(app)
-		if err != nil {
-			return fmt.Errorf("error getting %s release: %s", app, err)
-		}
-		formation, err := client.GetFormation(app, release.ID)
-		if err != nil {
-			// TODO: handle ErrNotFound
-			return fmt.Errorf("error getting %s formation: %s", app, err)
-		}
-		if app == "controller" {
-			controllerFormation = formation
-		}
-		for typ := range release.Processes {
-			var want int
-			if app == "postgres" && typ == "postgres" && len(f.hosts) > 1 && formation.Processes[typ] < 3 {
-				want = 3
-			} else if formation.Processes[typ] < 1 {
-				want = 1
-			}
-			if want > 0 {
-				f.l.Info("found broken formation", "app", app, "process", typ)
-				if _, ok := changes[app]; !ok {
-					if formation.Processes == nil {
-						formation.Processes = make(map[string]int)
-					}
-					changes[app] = formation
-				}
-				changes[app].Processes[typ] = want
-			}
-		}
-	}
-
-	for app, formation := range changes {
-		f.l.Info("fixing broken formation", "app", app)
-		if err := client.PutFormation(formation); err != nil {
-			return fmt.Errorf("error putting %s formation: %s", app, err)
-		}
-	}
-
-	if startScheduler {
-		if err := f.StartScheduler(client, controllerFormation); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (f *ClusterFixer) StartScheduler(client controller.Client, cf *ct.Formation) error {
-	if _, err := discoverd.NewService("controller-scheduler").Leader(); err != nil && !discoverd.IsNotFound(err) {
-		return fmt.Errorf("error getting scheduler leader: %s", err)
-	} else if err == nil {
-		f.l.Info("scheduler looks up, moving on")
-		return nil
-	}
-	f.l.Info("scheduler is not up, attempting to fix")
-
-	// start scheduler
-	ef, err := utils.ExpandFormation(client, cf)
-	if err != nil {
-		return err
-	}
-	schedulerJob := utils.JobConfig(ef, "scheduler", f.hosts[0].ID(), "")
-	if err := f.hosts[0].AddJob(schedulerJob); err != nil {
-		return fmt.Errorf("error starting scheduler job on %s: %s", f.hosts[0].ID(), err)
-	}
-	f.l.Info("started scheduler job")
-	return nil
-}
-
-func (f *ClusterFixer) KillSchedulers() error {
-	f.l.Info("killing any running schedulers to prevent interference")
-	for _, h := range f.hosts {
-		jobs, err := h.ListJobs()
-		if err != nil {
-			return fmt.Errorf("error listing jobs from %s: %s", h.ID(), err)
-		}
-		for _, j := range jobs {
-			if j.Job.Metadata["flynn-controller.app_name"] != "controller" || j.Job.Metadata["flynn-controller.type"] != "scheduler" {
-				continue
-			}
-			if j.Status != host.StatusRunning && j.Status != host.StatusStarting {
-				continue
-			}
-			if err := h.StopJob(j.Job.ID); err != nil {
-				f.l.Error("error stopping scheduler job", "id", j.Job.ID, "error", err)
-			}
-			f.l.Info("stopped scheduler instance", "job.id", j.Job.ID)
-		}
-	}
-	return nil
-}
-
-func (f *ClusterFixer) FixSirenia(svc string) error {
-	log := f.l.New("fn", "FixSirenia", "service", svc)
-	log.Info("checking sirenia cluster status")
-	service := discoverd.NewService(svc)
-	leader, _ := service.Leader()
-	if leader == nil || leader.Addr == "" {
-		log.Info("no running leader")
-		leader = nil
-	} else {
-		log.Info("found running leader")
-	}
-	instances, _ := service.Instances()
-	log.Info("found running instances", "count", len(instances))
-
-	log.Info("getting sirenia status")
-	var status *sirenia.Status
-	if leader != nil && leader.Addr != "" {
-		client := sirenia.NewClient(leader.Addr)
-		var err error
-		status, err = client.Status()
-		if err != nil {
-			log.Error("error getting status from leader", "error", err)
-		}
-	}
-	if status != nil && status.Database != nil && status.Database.ReadWrite {
-		log.Info("cluster claims to be read-write")
-		return nil
-	}
-
-	log.Info("getting service metadata")
-	meta, err := discoverd.NewService(svc).GetMeta()
-	if err != nil {
-		return fmt.Errorf("error getting sirenia state from discoverd: %s", err)
-	}
-
-	var state state.State
-	if err := json.Unmarshal(meta.Data, &state); err != nil {
-		return fmt.Errorf("error decoding state: %s", err)
-	}
-	if state.Primary == nil {
-		return fmt.Errorf("no primary in sirenia state")
-	}
-
-	log.Info("getting primary job info", "job.id", state.Primary.Meta["FLYNN_JOB_ID"])
-	job, host, err := f.GetJob(state.Primary.Meta["FLYNN_JOB_ID"])
-	if err != nil {
-		if state.Sync != nil {
-			f.l.Error("unable to get primary job info", "error", err)
-			f.l.Info("getting sync job info", "job.id", state.Sync.Meta["FLYNN_JOB_ID"])
-			job, host, err = f.GetJob(state.Sync.Meta["FLYNN_JOB_ID"])
-			if err != nil {
-				return fmt.Errorf("unable to get primary or sync job details: %s", err)
-			}
-		} else {
-			return fmt.Errorf("unable to get primary job details: %s", err)
-		}
-	}
-
-	if leader != nil && state.Singleton {
-		return fmt.Errorf("sirenia leader is running in singleton mode, unable to fix")
-	}
-
-	waitForInstance := func(jobID string) (func() (string, error), error) {
-		watchCh := make(chan *discoverd.Event)
-		upCh := make(chan string)
-		stream, err := service.Watch(watchCh)
-		if err != nil {
-			return nil, fmt.Errorf("error watching discoverd service: %s", err)
-		}
-		go func() {
-			var current bool
-			for event := range watchCh {
-				if event.Kind == discoverd.EventKindCurrent {
-					current = true
-					continue
-				}
-				if !current || event.Kind != discoverd.EventKindUp {
-					continue
-				}
-				if event.Instance.Meta["FLYNN_JOB_ID"] == jobID {
-					upCh <- event.Instance.Addr
-				}
-			}
-		}()
-		return func() (string, error) {
-			log.Info("waiting for instance to start", "job.id", jobID)
-			defer stream.Close()
-			select {
-			case addr := <-upCh:
-				return addr, nil
-			case <-time.After(time.Minute):
-				return "", fmt.Errorf("timed out waiting for sirenia instance to come up")
-			}
-		}, nil
-	}
-
-	var wait func() (string, error)
-	have := len(instances)
-	want := 2
-	if state.Singleton {
-		want = 1
-	}
-	if have >= want {
-		return fmt.Errorf("already have enough instances, unable to fix")
-	}
-	log.Info("attempting to start missing jobs", "want", want, "have", have)
-	if leader == nil {
-		job.ID = cluster.GenerateJobID(host.ID(), "")
-		f.FixJobEnv(job)
-		log.Info("starting primary job", "job.id", job.ID)
-		wait, err = waitForInstance(job.ID)
-		if err != nil {
-			return err
-		}
-		if err := host.AddJob(job); err != nil {
-			return fmt.Errorf("error starting primary job on %s: %s", host.ID(), err)
-		}
-		have++
-	}
-	if want > have {
-		// if not enough postgres instances, start another
-		var secondHost *cluster.Host
-		for _, h := range f.hosts {
-			if h.ID() != host.ID() {
-				secondHost = h
-				break
-			}
-		}
-		if secondHost == nil {
-			// if there are no other hosts, use the same one we put the primary on
-			secondHost = host
-		}
-		job.ID = cluster.GenerateJobID(secondHost.ID(), "")
-		f.FixJobEnv(job)
-		log.Info("starting second job", "job.id", job.ID)
-		if wait == nil {
-			wait, err = waitForInstance(job.ID)
-			if err != nil {
-				return err
-			}
-		}
-		if err := utils.ProvisionVolume(secondHost, job); err != nil {
-			return fmt.Errorf("error creating volume on %s: %s", secondHost.ID(), err)
-		}
-		if err := secondHost.AddJob(job); err != nil {
-			return fmt.Errorf("error starting additional job on %s: %s", secondHost.ID(), err)
-		}
-	}
-
-	if wait != nil {
-		addr, err := wait()
-		if err != nil {
-			return err
-		}
-		if leader != nil {
-			addr = leader.Addr
-		}
-		log.Info("waiting for cluster to come up read-write")
-		return sirenia.NewClient(addr).WaitForReadWrite(5 * time.Minute)
-	}
-	return nil
 }
 
 func (f *ClusterFixer) GetJob(jobID string) (*host.Job, *cluster.Host, error) {
